% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ parsetR }
\alias{ parsetR }
\alias{ parsetR }
\title{ Parallel Sets (d3.js) in R }
\description{
Visualize your data with interactive d3.js parallel sets with the power and convenience of an htmlwidget.
}
