% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parset.R
\name{parset}
\alias{parset}
\title{Create a Parallel Sets Visualization}
\usage{
parset(data = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{}
}
\description{
Create a Parallel Sets Visualization
}

